validators/cli/python/pyproject.toml
[project]
name = "universal-spawn"
version = "1.0.0"
description = "Validator + CLI for the universal-spawn manifest standard."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
  { name = "Jani Solo", email = "jani@agentmindcloud.dev" },
]
keywords = ["universal-spawn", "manifest", "json-schema", "spawn"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "jsonschema>=4.20",
  "pyyaml>=6.0",
  "referencing>=0.30",
]

[project.scripts]
universal-spawn = "universal_spawn.cli:main"

[project.urls]
Homepage = "https://universal-spawn.dev"
Repository = "https://github.com/AgentMindCloud/universal-spawn"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/universal_spawn"]
include = ["src/universal_spawn/_bundled/**/*.json"]

[tool.hatch.build.targets.sdist]
include = ["src/", "tests/", "README.md", "pyproject.toml"]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
